Huge pile of changes that makes the code up to date with our ideas.
The main thing is now a strong separation between putting an op into the ops
data structure, and the subsequent execution of it. Also, how the initial
cursors end up in the ops data structure is much different, but probably more
in line with what we're actually designing. I added the status register and a
mutation phase register too. The mutation phase register points to a set of
cursors that the current operations should use if they have side effects on the
scene tree and stuff. I tried to reify the "cursor context API" more too and
was mostly successful.

There is an interesting interplay between FC, CC, and MC registers that
we can exploit.

I could have refactored even more, but I don't want to unless we think it
is necessary since it begins to hide the ideas at that compression level of
the code.