Move op/clear-mutation-phase to be placed during set-mutation-phase.
How a mutation phase automatically and fully generates, and how
it knows to STOP generation is something that needs solidifcation.
I think we should have a ope before and after the mutation phase which
sets a flag in the SR and the ops twiddle the flag. If the bounding
op to the mutation phase sees the op was twiddled, it meant another
mutation phase is likely so generate it and place it into the frame
using the frame cursor contexts. But, if a mutation phase passed with
the flag NOT being twiddled, then the mutation phase is over, all mutations
are complete.