Finally figured out a linear extension for multi column sorting.
The big idea is that when a sorting class defines named arguments
it must be the fact they are also in an order. All derived class
are FORBIDDEN to alter that order, but they may stick new ordered
columns to the left or right of the previous columns. But, once
those columns are fixated, all children classes must adhere to it.

All components have a base sorting class that contains
@ <class-type> <instance-id> and a gamedev can only put new argument at the
@ sign.